Split sync code into multiple files
